WEBBPMNEXT-9355: move custom exception to separate module
