+++ jbpm-flow/src/main/java/org/jbpm/workflow/core/node/Split.java
@@ -19,7 +19,7 @@ package org.jbpm.workflow.core.node;
 import java.util.Collections;
 import java.util.Map;
 
-import org.jbpm.bpmn2.BpmnNodeIllegalArgumentException;
+import ru.micord.webbpm.jbpm.validation.exception.BpmnNodeIllegalArgumentException;
 import org.jbpm.workflow.core.Constraint;
 import org.jbpm.workflow.core.impl.ConnectionRef;
 import org.jbpm.workflow.core.impl.NodeImpl;
